LayoutTests/accessibility/mac/aria-modal-auto-focus-expected.txt

This tests that focus will automatically move to aria-modal dialogs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS backgroundAccessible() is true
PASS newBtn.isFocused is true
PASS closeBtn.isFocused is true
PASS okBtn.isFocused is true
PASS successfullyParsed is true

TEST COMPLETE
Other page content with a dummy focusable element

Display a dialog

Just an example.

OK  New
